vpython: Consolidate initialization logic.
Consolidate initialization logic paths to remove the no-wheel
environment one-off initialization and allow for a fully lockless
initialization.

Previously, a special "load exactly what you need" initialization path
was employed for environments with wheels to load their PEP425 tags from
an empty minimal environment. That environment wasn't being credited
with its usage and was not being exempted from pruning. In practice,
this is probably not a big deal, since pruning timestamps would be large
enough for this to be a non-issue. However, we should acknowledge it.

BUG=None
TEST=None

Review-Url: https://codereview.chromium.org/2847403002