Milo flex raw log viewer endpoint
This is a simple flex app that runs the logdog fetcher and serves back log text.  It runs in flex to avoid the GAE classic restrictions such as response size and response time.

Most of the logic is setting up a flex environment, most notably, memcache is not yet available.  The idea is to refactor this to look more like the GAE Classic middleware once a public memcache API is released.  Until then, memcache is simply not used.

BUG=698429

Review-Url: https://codereview.chromium.org/2796743004