Update LOADING status after a while.
After a log stream has been loading for a while, update its status to
note that the build may have crashed.

This includes the addition of a Timer class, which can be used to start
and stop a "setTimeout"-based timer. The sleep Promise has been been
updated to use the Timer.

BUG=chromium:703009
TEST=local

Review-Url: https://codereview.chromium.org/2855923003