Enable sequential prefix index logs to be loaded.
The LogDog viewer requires at least one log from each muxing stream to
know which one is sequentially next. For timestamp-based comparison, this is
a hard requirement, since the next log in any given stream could be
sequentially next. However, for prefix index-based comparison, the
viewer could know that a log from another buffer is next if its prefix
index is contiguous.

Enable this inference in the viewer: if a log stream buffer holds a
prefix index that is one more than the last buffered log, allow it to be
considered as the next log, even if other buffers are empty.

BUG=chromium:703428
TEST=prod

Review-Url: https://codereview.chromium.org/2861583003