vpython: handle vpython vs. Python flags better.
Currently, "vpython" has trouble differentiating between flags meant for
it (e.g, "-log-level") and flags meant for Python. Since the two are
non-intersecting sets, we should be able to do something here.

This patch makes it so that "vpython" arguments MUST come first. It
scans forwards from the beginning and identifies and consumes any flags that
belong to its application FlagSet. If it encounters any flags or
aguments that don't, or if it hits a divider, it passes the remainder to
its subprocess.

BUG=None
TEST=local

Review-Url: https://codereview.chromium.org/2863743003