vpython: Hold VirtualEnv lock during execution.
Currently, "vpython" marks its presence known by periodically updating a
completion flag in a goroutine. This is prone to some unlikely race and
failure conditions.

This patch modifies this behavior. Instead of periodically updating a
flag, a "vpython" instance will hold a shared VirtualEnv lock throughout
its execution to indicate its utility. The environment completion flag
timestamp will be updated at the beginning end of execution.

This ensures that the "vpython" instance is represented throughout its
lifecycle. Other "vpython" instances that want to prune the environment
must acquire an exclusive lock, so no timing or configuration races can
occur.

The Delete method has been modified to require an exclusive lock. This
means that if any enviornment is being operated on, either in setup
(exclusive) or running Python (shared), the deletion will fail to obtain
the lock and will, therefore, fail. Previously, it was called from within a
"With" callback, which was always a bit sketchy and potentially unsafe.

BUG=None
TEST=local

Review-Url: https://codereview.chromium.org/2859973003