[logdog-stream-view] Fix loading loop.
Fix the log loading loop construction for LogDog stream view. The
current form included indefinite recursion and no clear terminal point.
This is an artifact of the transition to async notation.

Now, an operation is scoped to the lifecycle of its initiating request,
regardless of how many fetches it is composed of. The fetch loop is
clearly defined in a "for" loop that performs the fetch cycle repeatedly
until a terminal condition is hit.

The "been a while" timer is now fully scoped to a given fetch loop,
avoiding a bug where old timers would trigger during later cycles,
maintaining a near-perpetual state of "been a while" messaging
regardless of log throughput.

BUG=None
TEST=local
  - Ran in local browser against development instance, works.

Review-Url: https://codereview.chromium.org/2893783002