Introduce luci-go components version metric.
A large chunk of luci-go library is used as a common GAE runtime for many LUCI
and Chrome Infra apps. It becomes hard to track what apps are outdated and need
to be redeployed.

This CL makes each individual GAE process report 'luci/components/version'
metric with semantic version of the luci-go GAE library, so we can precisely see
what is currently running in production.

The code is a bit more general to eventually allow reporting versions of other
optional components (like Tumble and pRPC server runtime).

R=iannucci@chromium.org, dnj@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2870293002