[vpython] Verify environment, named installs.
Support installation of named enviornments, which are VirtualEnv whose
name is explicitly specified instead of implicitly inferred from the
supplied hash. This can be used by systems that want to use "vpython" to
bootstrap a VirtualEnv in a known place instead of default usage.

Since the specification is no longer always derived from the name, we
add a step to verify that a given VirtualEnv actally matches the
expected VirutalEnv. If it doesn't, it's considered incomplete and
reprovisioned. This allows a path to upgrade named "vpython" roots from
one installation to another, as well as additional cheap sanity checking
to the basic "vpython" assumption that a stamped VirtualEnv is complete.

BUG=None
TEST=local
  - Ran "install" and setup normally, appears to work.

R=sergeyberezin@chromium.org

Review-Url: https://codereview.chromium.org/2918623003