[vpython] Fix flag parsing bug with "--".
When flags are being parsed, if a flag intended for Python (not vpython)
is encountered, and a subsequent argument includes "--", the parser will
only return flags between the Python flag and the "--" delimiter, not
the "--" delimiter or any arguments following it. This is incorrect.

Fix the parser to return all arguments past the Python flag instead, and
add a test for this.

BUG=None
TEST=unit

Review-Url: https://codereview.chromium.org/2937123003