token-server: Make machine token minter config apply to subdomains.
Token server is configured with a list of domains it is allowed to mint machine
tokens for. This list used to be matched exactly, e.g a cert with Common Name
"machine.sub.example.com" can be exchanged to a machine token only if
"sub.example.com" is listed in the config.

This change relaxes this rule and allows configured domains to match subdomains
too.

Also cleanup some stale assumptions in comments and code.

R=smut@google.com, iannucci@chromium.org
BUG=732467

Review-Url: https://codereview.chromium.org/2952723003