[errors] de-specialize Transient in favor of Tags.
Previously 'transience' was a hard-coded attribute of the errors package. This
change adds 'tags' as a way to put attributes (potentially with some data) onto an
error as it goes up the stack.

This removes multiple symbols from the errors package (the stack context stuff
is now an implementation detail, tags are better than ExtractData).

The 'transience' bit has now moved to the 'common/retry/transient' package.
Marking an error as transient is now `err = transient.Tag.Apply(err)`, or
`errors.Annotate(err).Tag(transient.Tag).Err()` (as part of another batch
of annotations).

API Review in a Nutshell:
  // for simple 'boolean' tags, where presence is the only important piece of
  // information.
  var FatalTag = errors.BoolTag{errors.NewTagKey("this is a fatal error!")}

  func (...) error {
    return FatalTag.Apply(err)
  }

  func (...) {
    if FatalTag.In(err) {
      os.Exit(-1)
    }
  }

  // tags may also have associated typesafe values
  type Severity int
  type severityTagType struct {Key errors.TagKey}
  func (t severityTagType) With(s Severity) errors.TagValue { return errors.MkTagValue(t.Key, s) }
  func (t severityTagType) In(err error) (s Severity, ok bool) {
    d, ok := errors.TagValueIn(t.Key, err)
    if ok {
      s = d.(Severity)
    }
    return
  }
  var SeverityTag = severityTagType{errors.NewTagKey("holds a Severity value")}

  func (...) error {
    return SeverityTag.With(Severity(10)).Apply(err)
  }

  func (...) {
    if s, hasSeverity := SeverityTag.In(err); hasSeverity {
      if s.(Severity) > 9000 {
        fmt.Println("it's pretty severe")
      }
    }
  }

BUG=

Review-Url: https://codereview.chromium.org/2951393002