[errors] Greatly simplify common/errors package.
Now that we have tags with values, the original concept behind the 'data'
stuff in errors no longer applies. This removes the Reason().D() system and puts
the format directly into errors.Annotate. So:
  errors.Annotate(err).Reason("foo: %(datum)s").D("datum", "value").Err()

Becomes:
  errors.Annotate(err, "foo %s", "value").Err()

This also removes a bunch of rarely-or-never-used symbols from common/errors by
making them internal (particularly w.r.t. rendering an error). If we ever develop
a need for them, it would be easy to bring them back.

R=dnj@chromium.org, vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2963503003 .