[vpython] Enable base wheels, spec merging.
Enable base wheel support. This allows an application instance to
specify a set of wheels that will always be included with every vpython
spec, regardless of whether or not the spec explicitly specifies them.

To support this, enhance wheel merging. Duplicate wheel declarations which
share the same versions will have their parameters deduplicated and
merged into a single wheel spec.

BUG=chromium:740320
TEST=None

Review-Url: https://codereview.chromium.org/2971403002