[logdog] Begin migrating to Makefile.
"deploytool" didn't catch on, so begin migrating LogDog's AppEngine
provisioning to a Makefile. This Makefile will consist of two
components: the "luci-go" component, which is a general set of rules to
build, and a deployment component, resident in a separate directory,
which populates some constants and imports the main Makefile.

BUG=None
TEST=local

Review-Url: https://codereview.chromium.org/2988083002