[logdog-view] Update UX, fix bugs.
Update the LogDog Web Viewer user experience and fix a bug where the
stream finishing wasn't always acknowledged by the UI.

This makes button visibility modal. While the stream is automatically
loading, control buttons will be hidden. When it is paused, they will be
shown. This will reduce the number of active buttons that the user has
to see during log interaction.

Following is no longer optional. It is enabled by default, and disabled
permanently for the duration of the stream on scroll or pause. This
removes the need for a user to manage this option.

BUG=None
TEST=local

Review-Url: https://codereview.chromium.org/2988993003