[logdog] Replace Tumble with push queues.
Replace Tumble with push queues for archival.

Currently, LogDog uses the Tumble journaling state machine to manage its
archival. Tumble is generally overkill for this one-state task, but was
chosen because it seemed, at the time, likely that it would be used
everywhere in LUCI.

Almost two years later, LogDog is the only production major user of
Tumble. Since it barely scrapes the power of Tumble, and since Tumble
itself is rather opaque in its operations, this trade-off is not
worthwhile. Instead, we replace Tumble with task queues.

When a log stream is registered, an "expired" task will be enqueued to
handle it once the stream expires (if it never gets terminated). When
the stream is terminated, the expiration task is deleted, replaced with
a shorter-term archival task.

We leave Tumble and its mutation handling in-place because, in
production, there is still a Tumble backlog to process through. This
should be empited within a few days, and we can finish the removal.

BUG=chromium:751925
TEST=unit

Review-Url: https://codereview.chromium.org/2989333002