[lru] Update, notably expiration.
Update the LRU cache implementation to allow for item expiration. This,
in turn, updates the function signatures of most of the cache functions
to include a Context and introduces a Clock dependency.

We additionally update Mutate to operate on a special cache Item, since the
function signature was getting confusing and unruly.

This prepares LRU for wider use as a manageable alternative to other
expiring caches, notably memcache.

BUG=chromium:753846
TEST=unit