scheduler: make disabled jobs not visible.
Once a job is deleted from a config, it should be invisible for
everybody.

If project wants it visible but not firing, it's scheduled can be
changed "triggered", perhaps with a comment that job is disabled
but kept for posterity. That said, this is unlikely to be useful
for long time since most artifacts of the scheduler (ie. builds)
will be in buildbucket/swarming/etc anyway.

R=vadimsh@chromium.org

Bug: 736770
Change-Id: I3133c0bcf07c48f37f199789a5225247d88ddc7f