[server/auth] Expose MemoryCache publicly.
Make the MemoryCache in-memory LRU caching struct public. This allows it
to be selectively used outside of the auth package, as well as for
arbitrary LRU caches to be plugged into it.

The cache was previously used just for tests, but will see expanded use
in environments without built-in caches such as AppEngine Flex.

Riding on this, update the call signature of "fetchOrMintToken" to make
the error the last return value so that it passes linting checks.

BUG=chromium:753846
TEST=None

Change-Id: I8b35e351f07d0187de94ff1bd5b2682c3d7c17e3