[luci_config] Replace prooccache with LRU.
Add an LRU-based caching mechanism. This cache uses an underlying LRU
structure to store configurations. Because we're using an LRU instead of
a proccache, we can comfortably store individually-authenticated
configuration results in the cache as well.

Replace the proccache caching backend with an LRU-based cache. This adds
more versatility, as callers can supply LRUs with different constaints.

BUG=chromium:753846
TEST=unit

Change-Id: I679c783d0c13b6950ed6a2c6c5349161488036e1