Remove unneeded GetConfigSetURL.
GetConfigSetURL is only used for getting a display revision URL.
https://chromium-review.googlesource.com/c/608753 removes this use case (by
surfacing revision URL directly).

Add ViewURL to config metas.

Currently, the frontend displays a config URL that is constructed using config
location, revision, and config path, and a guess as to the structure. This guess
is not alway correct.

This CL augments the API with a config URL for the frontend to display directly
instead.

Bug:
Change-Id: I333cbd3a42f8f7992601e3fbe10e931ee14c84ea