isolatedclient: Use a buffer instead of a pipe for http.Request
procmon shows that isolated client calls socket.Write() for every 278 bytes,
which is fine on linux/mac if the kernel buffers the network calls, but causes
windows to send out packets that are sized ~300 bytes even though the MTU
is 1500.  This adds buffering between the zlib compressor and http client
so that it can read as many bytes as it wants.

BUG=753184

Change-Id: Ic17f5eb548086db5be62a323ade0f568d2adbc20