Add ViewURL to config metas.
Currently, the frontend displays a config URL that is constructed using config
location, revision, and config path, and a guess as to the structure. This guess
is not alway correct.

This CL augments the API with a config URL for the frontend to display directly
instead.

Bug: 747645
Change-Id: I1bbfcc8d052441b64d314ccd33715c9b69602262
Reviewed-on: https://chromium-review.googlesource.com/608753
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>