[server] Add request and global caches.
Add request and process-global object caching to context. These mechanisms
can be used by services and their support libraries to store information.

Request caching installs a cache into the Context that exists for the
lifetime of a single HTTP request. Global caching is a fixed-size LRU
that survives for the lifetime of the process.

Nothing uses these yet, but future CLs will migrate existing cache
usage to the new semantics.

BUG=chromium:753846
TEST=None

Change-Id: I1dc66516e73596d63e9b5e5dbea1b430b6c30435