[cipd] Move template expansion code to common.
This will enable us to expand packages in other contexts (such as
command-line parameters).

This also fixes subdir expansion to behave more like package name
expansion; it uses the expansion paramaters passed to ResolveWith,
and if a subdirectory uses the ${param=value} syntax and doesn't 
match the host, the entire subdirectory is skipped.

R=vadimsh@chromium.org

Bug: 756288
Change-Id: Ic9f9bf9fb0f8e64a45bc99c30d3a78f6c8acbb49