[gaemiddleware] Split classic environment.
Update the "gaemiddleware" package and its users to explicitly install a
"classic AppEngine" environment, and to leave room for other
environments (e.g., Flex) to be defined.

The "gaemiddleware" package is a starting point for application setup.
However, it is very specific to Classic AppEngine environments. Most of
the code in the package is precise and generally-useful, but it is
interleaved with AppEngine Classic requirements.

Update the package to define a generic Environment, which is a skeleton
for specific instantiations. Create a "classic" sub-package that
implements that skeleton for AppEngine Classic environments, and update
callers to use the "classic" package.

BUG=chromium:753846
TEST=None

Change-Id: Icfecfd8c65bcacdaa2ece73a330c286aa2361162