[gaemiddleware] Split standard environment.
Update the "gaemiddleware" package and its users to explicitly install a
"AppEngine Standard" environment, and to leave room for other
environments (e.g., Flex) to be defined.

The "gaemiddleware" package is a starting point for application setup.
However, it is very specific to Standard AppEngine environments. Most of
the code in the package is precise and generally-useful, but it is
interleaved with AppEngine Standard requirements.

Update the package to define a generic Environment, which is a skeleton
for specific instantiations. Create a "standard" sub-package that
implements that skeleton for AppEngine Standard environments, and update
callers to use the "standard" package.

BUG=chromium:753846
TEST=None