[lru] Fix GetOrCreate, add Create.
GetOrCreate required a lock on an item's key before it would check for
the item's existence. This breaks the case where an item is already in
the cache, but another coroutine chooses to generate a new one while the
item is still valid (e.g., preemptively refreshing tokens). As-is, the
"create" attempt would cause valid "get" to block, undermining the
purpose of this model.

Make "GetOrCreate" perform the "get" part of the call unlocked. This allows
a valid item to be returned.

Additionally, introduce "Create", which will always regenerate the item
under lock. If the "-OrCreate" part of "GetOrCreate" is hit, it will
serialize with Create.

BUG=chromium:753846
TEST=unit

Change-Id: I0a3c911153ab13cb6702cd2bd7b75dc56f0f6f75