[luci_config] Add dedicated erroring interface.
The current "impl/memory" LUCI config interface is used mostly in
development, but also in production for its ability to produce a
consistent error. We want to make it used only in development. To do
this, we will add a simple config.Interface implementation, "erroring",
which will always return an error, and use it instead of "impl/memory".

BUG=None
TEST=unit

Change-Id: I32aba6503a2a44c293b322dd21004ea1f520492b