isolate: move logging into doExpArchive and doArchive
This reduces the amount of code which is specific to archiveRun.main vs
expArchiveRun.main. This means that once archiveRun.main starts
optionally calling doExpArchive based on a flag, there won't be a lot of
code which varies between the two branches.

This also has the effect of making the exparchive output more closely
match the archive output, and removing the need to pass "quiet" and
"start" into doArchive.

Bug: 692940
Change-Id: Ibaf4afe09ab8cf9775c90a93c502b641a8bf40cc