Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Missing tests; will add to routes_test and unit tests for the tree structure in api/resp/console.go. I'm unsure about whether I should keep so much logic in api/resp or whether that should be moved, also.

I forgot to mention that there's also some inefficiencies: for example I have both a map and an array in Category in api/resp/console.go. In addition to that, the map in Category contains strings which are the same as those found in the BuilderRef and Category structs' Name fields (I get that the golang compiler likely canonicalizes them, but it's still an extra pointer at least). I wanted some input on the best way to resolve this or whether to just leave it the way it is for now.

Patch-set: 1
Reviewer: Michael Knyszek <1154295@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0