[logdog/fetcher] Add option to early-bail on incomplete streams.
This is useful in applications where the user of Fetcher expects the
stream to already be completed, but hasn't verified this with LogDog
yet. Without this option, the user would block until the context was
Canceled, without any feedback that the stream was incomplete.

This is more efficient and more ergonomic than making a separate RPC
to LogDog, since Fetcher already has the the necessary guts to determine
this. It's especially ergonomic when used in conjunction with the
Reader() API.

R=dnj@chromium.org

Bug:
Change-Id: Ie8494031d3e4f28e3f155801949681d85450e55a
Reviewed-on: https://chromium-review.googlesource.com/642716
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Daniel Jacques <dnj@chromium.org>