[scheduler] teach engine to propagate triggers.
This CL end-goal is gitiles-based triggering of buildbucket jobs.
It builds on top of state machine processing of triggers and wires
trigger passing through the engine, such that a trigger can be
sent from context of one job execution to other jobs.

This CL does NOT actually implement usage of triggers.

Bug: 646067
Change-Id: I2ed2ad1d822a5f0f68e3cfad8b7ac10a95bc76ac