scheduler: Add jobID argument to GetVisibleInvocationsByNonce.
This slightly simplifies switching to v2 jobs since we'll know in advance what
kind of invocations to expect from the query (v1 or v2), based on the known
job ID.

This is also slightly safer, since we know guarantee that caller sees only
invocations for the job they requested. Before there was (very small) chance
to encounter a collision in nonces and see someone else's invocations.

R=tandrii@chromium.org
BUG=764043

Change-Id: I14bba2ad8124e313b3be352db23672a6b3978004