[vpython] Use execve on Linux/Mac.
Use "execve" to launch Python on Linux/Mac. This causes "vpython" to
*become* the target Python rather than simply host it, increasing the
accuracy of the wrapper's Python-like behavior.

We want to be non-invasive, meaning we don't want to modify the target
Python, including injecting the shared lock file descriptor.
Accordingly, we spawn a monitor process alongside the Python environment
to ensure that its lock is held. This monitor process holds its own
version of the lock and periodically polls the Python process to see if
it exists. If it does not, the monitor process terminates.

BUG=None
TEST=local

Change-Id: I0cedeb22e703f5206a7fe70ac8b761d29b80046f