Update patch set 2
Patch Set 2:

(2 comments)

Seems fine. Alternatively, we could have the spec file explicitly mention the platforms to skip. This would make that list explicit, with the opportunity to comment on why, and remove the need for the environment variable.

Sure, every spec would have to mention this, but maybe they should. Maybe a platform that we can't or don't build a wheel for is an exemption in one situation and not in another. Look into this maybe? spec.proto.

Patch-set: 2
Reviewer: Daniel Jacques <1109829@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>