Update patch set 2
Patch Set 2: Code-Review+1

> I'm less worried about that at this point; managing these things is so hard, and the design of this envvar should give the user a feeling that they're going deep into uncharted territory. I can make the text of the envvar even more strict:
> 
>   VPYTHON_BYPASS="I expect demons to fly out of my nose and promise not to file bugs"

The text is fine for the direction we're going.

I'm still inclined to say that for *our* version of the tool, we should make it explicitly baked into the tool so we can manage it. If someone else wants to maintain a list independently from us, they're free to fork "infra/tools/vpython" and add this or some other feature. We're already sort of in this bind, since we're now building this and other tools for platforms we don't support at all. Ideally the other team would build their own version of this, and we could let them override *that* with an env var maybe in the "vpython" launch script.

OTOH I understand the appeal and rationale behind providing an open escape hatch. I'm not wholly opposed to it, but I do think it invites some unappealing scenarios:

* Trooper sees a problem w/ vpython or related library. Is breaking builds, P0. Rather than look into it, they see escape hatch, commit recipe change, and builds work again. It seems like they've solved the problem, but really they just cast it into the wind.
* User encounters a situation where a wheel is not built for their platform. Rather than build it, they just locally disable vpython for that part of their recipe via environment.

It's not laziness; it's that we have people with different understandings and priorities, and a formal mechanism like this may seem like the right mechanism to use to someone else. I'm not totally happy empowering people who are not primary owners of this tool to make that call.

That said, since you and vadimsh@ are both on board with this approach, I don't want to hold it up. I can't think of a better way to empower users to disable the tool than this approach, so I'll LGTM.

Patch-set: 2
Reviewer: Daniel Jacques <1109829@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1