Update patch set 3
Patch Set 3:

(1 comment)

The single row preview looks nice, but it would be more useful to get the latest completed status rather than the status of the last build.  It's find for now though.

Also it looks like the bottom of the table is missing a border.  I guess it was always missing, but the single row makes it more obvious.

Patch-set: 3
Reviewer: Ryan Tseng <1002080@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>