[logdog] Move Coordinator frontend over to Flex.
Switch the LogDog Coordinator frontend from GAE Classic to GAE Flex.
This includes refactoring of some of its clients to be global and
movement of its endpoints into a new "flex/" package with "flex"-style
Services.

Because Flex functions, in many ways, fundamentally differently than
classic, it is necessary to draw boundaries between code intended to
operate in a Flex environment and code for a GAE Classic environment.
Because Flex allows connections to persist outside of the scope
of individual HTTP request handling, a Flex enviornment is generally
simpler.

BUG=chromium:753846
TEST=unit

Change-Id: Iafe0244ba9ead05c7be191049f0303f76b9ec290