[vpython] Use execve on Linux/Mac.
Use "execve" to launch Python on Linux/Mac. This causes "vpython" to
*become* the target Python rather than simply host it, increasing the
accuracy of the wrapper's Python-like behavior.

The VirtualEnv lock file descriptor is passed to the spawned process
as file descriptor #3. This modifies the Python runtime, as FD #3 is
now in-use and reserved. However, this is fairly non-invasive for
both POSIX and Python invocation expectations. It has the bonus of
implicitly passing lock ownership to daemoized and grandchild
processes.

We considered spawning a monitor process instead in order to be
non-invasive, but this carries its own issues, including doubling the
process count and imperfect monitoring and complexity.

Lock persistence is non-critical, and an unlocked VirtualEnv is
probably not going to cause impact when deleted after delay, since
its host program will have run for a while and its imports will
probably be complete. We're OK with best-effort for now.

BUG=None
TEST=local

Change-Id: I0cedeb22e703f5206a7fe70ac8b761d29b80046f