Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> High level comment: I would prefer that this new struct be defined as proto and have a couple convenience methods in the generate proto package to do proto.Timestamp -> time.Time, etc.
> 
> The reasons for this would be:
>   * We could use this as an exact representation of the v2 (g/prpc) api; if you convert your code to use this proto, you're already compatible with the v2 api
>   * If something about this struct format is awkward to represent in proto, we'll find out now, rather than later; this means that we can change this struct to be more homogenous with the proto way now, so that code can also adjust to this now. Otherwise we'll be potentially stuck with a complicated bi-directional conversion function (proto<->struct), or we'll end up ditching this struct anyway.
>   * Protos are nice data-only documentation of an interface. I personally find them really easy to read compared to JSON-in-Go-struct, or python endpoints formats.
>   * Protos are cool 😜
> 
> If we decide to go with the struct instead of a proto, I'll take a closer look at the particulars here.

I considered it to be a proto, but made it a more idiomatic struct because the problem I was trying to solve is to remove the need to parse stuff. With a proto, they will still have to parse stuff, but less. But I don't mind. With a proto we'd indeed simplify the transition to new APIs.

We already have time convenience methods: https://chromium.googlesource.com/infra/luci/luci-go/+/master/common/proto/google/util.go#53
I am not sure why it is a function and not a method though. I will fix that.

Patch-set: 5
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>