scheduler: implement memory efficient gitiles refs fetching.
Previously, just one Refs API was issued for **all** refs in a repo.
For Gerrit repos with lots of CLs in them like Chromium main repo,
this results in 10MB of just raw json string AND it will only grow.

This CL instead fetches tips of sub-namespaces such as "refs/heads"
and "refs/branch-heads", which may mean more requests but less
data returned in each request.

R=vadimsh@chromium.org

Bug: 767222
Change-Id: Iee278aa89201395f7ba86b01a22fb32559dd9451