auth: Add GetEmail method to Authenticator.
It returns an email associated with the token. It stores this email alongside
the token in the cache to avoid doing round trips to the token info endpoint
all the time.

It also uses different strategies for grabbing this email for different auth
providers. For example, on GCE we just ask the metadata server, which is faster
and more reliable than using token info endpoint.

LUCI_CONTEXT-based method for grabbing an email is not implemented yet. It will
be done in a separate CL (by storing the email in local_auth struct).

Implementation consists of replacing *oauth2.Token with *internal.Token
everywhere inside auth library guts, where internal.Token is just a pair
(oauth.Token, email). Public API still operates in terms of *oauth2.Tokens.

R=iannucci@chromium.org
BUG=767730

Change-Id: I350f9244521ec645ddc1d3376322943831114d95