auth: Avoid doing RPCs when getting an email, if possible.
When building a LUCI_CONTEXT with emails we need to get an email before we
actually get the token (email is needed to populate the environment, and the
token is needed only later when clients make RPCs).

Optimize this case to avoid any RPCs, if possible. This is not possible only
when using User provider (the one with manual login), and there's no token
cached, or it is in old format (without the email).

R=iannucci@chromium.org
BUG=767730

Change-Id: I73df79c259b0331ea76ff4c6d1a3a17d837d62ac