Update patch set 1
Patch Set 1:

PTAL

In previous CL I've implemented the code path that always works, but which is kind of heavy (it refreshes a token as a side effect of getting an email). This CL adds a short cut that makes not RPCs and works for all providers other than "User". This is important on bots.

Patch-set: 1
Reviewer: Vadim Shtayura <1002539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false