[buildbucket] add an idiomatic Build type
All *buildbucket.ApiCommonBuildMessage users have to do a lot of parsing
because it uses int64 for times, JSON-encoded fields for properties,
string for identity and []string for tags.

Add a more convenient idiomatic buildbucket.Build type.
The goal is to be as Go-native as possible.

Add (*Build).ParseMessage that converts *buildbucket.ApiCommonBuildMessage
to buildbucket.Build.

If this successful, we can design buildbucket API v2 based on this.

Bug:
Change-Id: I0d889972453d011da00cea5c9ee03a8fdf87faa9