Update patch set 16
Patch Set 16:

> Patch Set 5:
> 
> High level comment: I would prefer that this new struct be defined as proto and have a couple convenience methods in the generate proto package to do proto.Timestamp -> time.Time, etc.
> 
> The reasons for this would be:
>   * We could use this as an exact representation of the v2 (g/prpc) api; if you convert your code to use this proto, you're already compatible with the v2 api

yes, we could use this exact representation for v2 API.
It is not a reason not to have a nice idiomatic type, though.
I've updated CL description with the list of problems that this
CL attempts to solve. A proto-based type wouldn't solve them. Here is google-provided timestamp conversion functions 
https://godoc.org/github.com/golang/protobuf/ptypes#Timestamp
https://godoc.org/github.com/golang/protobuf/ptypes#TimestampProto
they return errors! which means one would have several `if err !=nil` lines instead of `just parse everything and tell me if there is a problem`.

>   * If something about this struct format is awkward to represent in proto, we'll find out now, rather than later; this means that we can change this struct to be more homogenous with the proto way now, so that code can also adjust to this now. Otherwise we'll be potentially stuck with a complicated bi-directional conversion function (proto<->struct), or we'll end up ditching this struct anyway.

I think this struct will be the Go API to Buildbucket because it would be nicer than any wire format and thus won't be ditched. Take a look at ParseMessage and PutRequest. I think protov2<->struct won't be more complicated than that. I think it is more important that buildbucket Go users have nice API than the buildbucket package has to do some conversion (for everyone's benefit).

>   * Protos are nice data-only documentation of an interface. I personally find them really easy to read compared to JSON-in-Go-struct, or python endpoints formats.

Note that this is not a JSON-In-Go-struct. It is intended to use native types wherever possible.

WDYT?

Patch-set: 16
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>