[milo] Refactor source manifest proto location and diff support.
Since the source manifests are a mechanism of communicating with milo, it
makes more sense to put them along-side the annotation proto definition,
not off on their own.

Additionally, since some services (like findit) are going to need to
consume the diff information in whole or in part, it also makes sense to
make the diff structure externally supplied as a proto too.

The initial diff implementation was also too simplistic to be useful to
callers, so I've embellished it somewhat, and moved it to live along-side
the protos themselves, so that Go programs can locally compute the
diff without importing any code from Milo, or being forced to make an
RPC.

R=hinoka@chromium.org, nodir@chromium.org

Bug: 758007
Change-Id: Ic2cf3f9af70b4d8ef1dd51e8ba40c0d184877c07